Fix: when adding attachments, check if parent index has changed (#189)
#### What problem is this solving?

When adding an item to the cart with multiple attachments, the existing
logic assumed that the index of the parent item would not change during
the operation. Normally, this is true. However, if you add item A with
attachments B,C,D to the cart and then add item E also with attachments
B,C,D, item E's index will change after each attachment is added. This
is because the cart will group attachments B,C, and D together. This
behavior is at the API level. In other words, when the second instance
of attachment B is added, even though it is attached to item E, it will
be inserted at the index immediately after the original instance of
attachment B.

This PR fixes the problem by re-checking the index of the parent item
after each attachment is added.

#### How should this be manually tested?

New version linked here: https://arthur--colgateb2b.mytex.com
Check this thread for how to test the scenario:
https://vtex.slack.com/archives/C02VB231E1Y/p1686787406485139

#### Checklist/Reminders

- [ ] Updated `README.md`.
- [x] Updated `CHANGELOG.md`.
- [ ] Linked this PR to a Jira story (if applicable).
- [ ] Updated/created tests (important for bug fixes).
- [ ] Deleted the workspace after merging this PR (if applicable).

#### Type of changes

<!--- Add a ✔️ where applicable -->
✔️ | Type of Change
---|---
✔️ | Bug fix <!-- a non-breaking change which fixes an issue -->
_ | New feature <!-- a non-breaking change which adds functionality -->
_ | Breaking change <!-- fix or feature that would cause existing
functionality to change -->
_ | Technical improvements <!-- chores, refactors and overall reduction
of technical debt -->